Began splitting MobileAppService into multiple classes Added support for "Scope" in Groups, where you can get more detail depending on the view requested (Coach, Member, NonMember)
