+++ church.ccv.CCVRest/MobileApp/Model/MAMessageModel .cs
