+++ church.ccv.CCVRest/church.ccv.CCVRest.csproj
@@ -99,26 +99,26 @@
     <Compile Include="Common\Model\PersonPhotoModel.cs" />
     <Compile Include="Common\Model\ResponseModel.cs" />
     <Compile Include="Common\Util\Util.cs" />
-    <Compile Include="MobileApp\Controllers\AccountMAController.cs" />
-    <Compile Include="MobileApp\Controllers\ContentMAController.cs" />
-    <Compile Include="MobileApp\Controllers\GroupMAController.cs" />
-    <Compile Include="MobileApp\Controllers\PersonMAController.cs" />
-    <Compile Include="MobileApp\Controllers\PodcastMAController.cs" />
+    <Compile Include="MobileApp\Controllers\MAAccountController.cs" />
+    <Compile Include="MobileApp\Controllers\MAContentController.cs" />
+    <Compile Include="MobileApp\Controllers\MAGroupController.cs" />
+    <Compile Include="MobileApp\Controllers\MAPersonController.cs" />
+    <Compile Include="MobileApp\Controllers\MAPodcastController.cs" />
     <Compile Include="MobileApp\Model\BaptismModel.cs" />
     <Compile Include="MobileApp\Model\CampusModel.cs" />
     <Compile Include="MobileApp\Model\CCVLiveCountdownModel.cs" />
-    <Compile Include="MobileApp\Model\FamilyMemberModel.cs" />
     <Compile Include="MobileApp\Model\JoinGroupModel.cs" />
     <Compile Include="MobileApp\Model\KidsContentModel.cs" />
-    <Compile Include="MobileApp\Model\MobileAppGroupModel.cs" />
+    <Compile Include="MobileApp\Model\MAGroupModel.cs" />
     <Compile Include="MobileApp\Model\LoginModel.cs" />
-    <Compile Include="MobileApp\Model\MobileAppMessageModel .cs" />
-    <Compile Include="MobileApp\Model\MobileAppPersonModel.cs" />
+    <Compile Include="MobileApp\Model\MAMessageModel .cs" />
+    <Compile Include="MobileApp\Model\MAPersonModel.cs" />
     <Compile Include="MobileApp\Model\NewUserModel.cs" />
     <Compile Include="MobileApp\Model\IsOnCampusModel.cs" />
     <Compile Include="MobileApp\Model\PersonalizedItemModel.cs" />
-    <Compile Include="MobileApp\Model\MobileAppSeriesModel.cs" />
+    <Compile Include="MobileApp\Model\MASeriesModel.cs" />
     <Compile Include="MobileApp\Model\PromotionModel.cs" />
+    <Compile Include="MobileApp\Util\MAGroupService.cs" />
     <Compile Include="MobileApp\Util\MobileAppService.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
